refactor(trusty-common): split payload_store.rs into submodules under 500-SLOC cap (#607)
Replaces the 547-SLOC payload_store.rs monolith with a payload_store/
directory module:

  payload_store/
    mod.rs    — module doc, re-exports, all tests (131 SLOC)
    types.rs  — PayloadStoreError, PayloadRow, PayloadRecord, Result alias,
                RowBytes, resolve_redb_path, decode_row, split_payload_key*
                (124 SLOC)
    store.rs  — PayloadStore struct + open/upsert/get/exists/delete/
                list_segment/load_all/iter_segment (305 SLOC)

Zero API change: PayloadStore, PayloadRow, PayloadStoreError remain pub
at the same path (memory_core::store::payload_store).
